@@ -1,21 +1,24 @@
-import { Controller } from '@nestjs/common';
-import { Ctx, MessagePattern, Payload, RmqContext } from '@nestjs/microservices';
+import { Controller, UseFilters } from '@nestjs/common';
+import { Ctx, MessagePattern, Payload, RmqContext, Transport } from '@nestjs/microservices';
 import { AuthService } from '../services/auth.service';
 import { UserAuth } from '../models/user.model';
+import { Observable } from 'rxjs';
+import { RpcPatternFilter } from '../filters/rpc-pattern.filter';
 
 @Controller()
+@UseFilters(RpcPatternFilter)
 export class AuthMessageController {
   constructor(private readonly authService: AuthService) {}
 
-  @MessagePattern('token-verify')
-  async findUser(@Payload() token: string, @Ctx() context: RmqContext): Promise<UserAuth> {
-    console.log('token-verify GET');
+  @MessagePattern('token-verify', Transport.RMQ)
+  findUser(@Payload() token: string, @Ctx() context: RmqContext): Observable<UserAuth> {
+    console.log('AuthService get message from "token-verify"');
     const channel = context.getChannelRef();
     const originalMsg = context.getMessage();
 
-    const user = await this.authService.verifyToken(token);
-
+    const observable = this.authService.verifyToken(token);
     channel.ack(originalMsg);
-    return 'user' as any;
+
+    return observable;
   }
 }
